backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
python-jose
passlib[bcrypt]
dnspython
